+++ R/is_model_supported.R
@@ -87,6 +87,7 @@ supported_models <- function() {
     "crqs",
     "crr",
     "DirichletRegModel",
+    "ergm",
     "feis",
     "felm",
     "feglm",