+++ R/format_p.R
@@ -21,6 +21,18 @@
 #' format_p(p, digits = "scientific")
 #' @export
 format_p <- function(p, stars = FALSE, stars_only = FALSE, name = "p", missing = "", digits = 3, ...) {
+
+  # only convert p if it's a valid numeric, or at least coercible to
+  # valid numeric values...
+  if (!is.numeric(p)) {
+    if (.is_numeric_character(p)) {
+      p <- .factor_to_numeric(p)
+    } else {
+      return(p)
+    }
+  }
+
+
   if (identical(stars, "only")) {
     stars <- TRUE
     stars_only <- TRUE